This commit was generated by cvs2svn to compensate for changes in r14247, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@14248 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/gfxoutputdrv/bmpdrv.c
@@ -225,7 +225,7 @@ static int bmpdrv_write(screenshot_t *screenshot)
           memset(sdata->bmp_data + (screenshot->height - 1 - sdata->line)
                  * screenshot->width/8, 0, screenshot->width/8);
 
-          for (i = 0; i < screenshot->width/8; i++)
+          for (i = 0; i < (int)screenshot->width/8; i++)
             {
               BYTE b=0;
               for (j = 0; j < 8; j++) b |= sdata->data[i*8+j] ? (1<<(7-j)) : 0;
